fix: Move key functions, fix bugs
- Improve overall functionality, fix bugs
- Move functions into __init__.py
- Improve overall documentation, comments, function header typing, and
code organization
